+++ pkg/local_object_storage/metabase/put_test.go
@@ -12,8 +12,6 @@ import (
 	cidtest "github.com/nspcc-dev/neofs-sdk-go/container/id/test"
 	objectSDK "github.com/nspcc-dev/neofs-sdk-go/object"
 	oidtest "github.com/nspcc-dev/neofs-sdk-go/object/id/test"
-	objecttest "github.com/nspcc-dev/neofs-sdk-go/object/test"
-	"github.com/stretchr/testify/require"
 )
 
 func prepareObjects(t testing.TB, n int) []*objectSDK.Object {
@@ -78,44 +76,5 @@ func BenchmarkPut(b *testing.B) {
 }
 
 func metaPut(db *meta.DB, obj *objectSDK.Object) error {
-	return db.Put(obj, nil)
-}
-
-func TestDB_PutBinary(t *testing.T) {
-	addr := oidtest.Address()
-
-	obj := objecttest.Object()
-	hdr := obj.CutPayload()
-	hdr.SetContainerID(addr.Container())
-	hdr.SetID(addr.Object())
-
-	obj = objecttest.Object()
-	hdrEnc := obj.CutPayload()
-	require.NotEqual(t, hdrEnc, hdr)
-	hdrEnc.SetContainerID(addr.Container())
-	hdrEnc.SetID(addr.Object())
-	hdrBin := hdrEnc.Marshal()
-	// although the distinction between a struct and a blob is not the correct
-	// usage, this is how we make the test meaningful. Otherwise, the test will pass
-	// even if implementation completely ignores the binary: header would be encoded
-	// dynamically and the parameter would have no effect. At the same time, for Get
-	// to work we need a match at the address.
-
-	db := newDB(t)
-
-	err := db.Put(hdr, hdrBin)
-	require.NoError(t, err)
-
-	res, err := metaGet(db, addr, false)
-	require.NoError(t, err)
-	require.Equal(t, hdrEnc, res) // exactly encoded
-
-	// now place some garbage
-	addr.SetObject(oidtest.ID())
-	hdr.SetID(addr.Object()) // to avoid 'already exists' outcome
-	err = db.Put(hdr, []byte("definitely not an object"))
-	require.NoError(t, err)
-
-	_, err = metaGet(db, addr, false)
-	require.Error(t, err)
+	return db.Put(obj)
 }